\name{alarm}
\docType{data}

\alias{alarmnet}

\title{The ALARM network}

\description{
  ALARM stands for 'A Logical Alarm Reduction Mechanism' and 
  it is a medical diagnostic alarm message system for patients monitoring developed by Beinlich et. all,  
  (Beinlich, I., Suermondth, G., Chavez, R., Cooper, G., The ALARM monitoring system, 
    1989, In Proc. 2-nd Euro. Conf. on AI and Medicine). 
  It is categorical Bayesian network has 37 nodes and 46 directed edges. 
}
\usage{data(alarmnet)}
\format{A data frame with 37 variables and 2000 samples.}
\source{http://www.norsys.com/netlib/alarm.htm}

\keyword{datasets}
