\name{as.graph-methods}
\docType{methods}
\alias{as.graph}
\alias{as.graph,catNetwork-method}

\title{Graph from catNetwork}

\description{Creates a \code{graph} object, as defined in \code{graph} package, from a \code{catNetwork} object.}

\usage{
as.graph(object)
}

\arguments{
\item{object}{a \code{catNetwork}}
}

\value{a \code{graph} object.}

\author{N. Balov}

\keyword{methods}
\keyword{graphs}
