\name{as.igraph-method}
\docType{methods}
\alias{as.igraph}
\alias{as.igraph,catNetwork-method}

\title{catNetwork to igraph}

\description{Create an \code{igraph} object from a \code{catNetwork}}

\usage{
as.igraph(object)
}
\arguments{
  \item{object}{a \code{catNetwork} object}
}

\value{An \code{igraph} object}

\author{N. Balov}

\keyword{methods}
\keyword{graphs}
