\name{breast}
\docType{data}
\alias{breast}

\title{Breast cancer data}
\description{
  Subclass Mapping: Identifying Common Subtypes in Independent Disease Data Sets
}
\usage{data(breast)}
\format{A matrix containing 100 observations on 1214 genes.}
\source{"http://www.broadinstitute.org/cgi-bin/cancer/datasets.cgi"}

\keyword{datasets}
