\name{classification}
\docType{data}
\alias{cvKfold}
\alias{diabetes}
\alias{boston}
\alias{lungLoad}
\alias{diabetesLoad}
\alias{diabetes}

\title{Classification demonstration}

\description{
  Detailed information on the analysis can be found in our paper "Discrete Bayesian Network Classification for Gene Expression Data". From the installation catnet/demo directory copy the files cvKforl.r, diabetesLoad.r, diabetes.r, bostonLoad.r and boston.r into a new directory along with the data files "Diabetes_collapsed_symbols.gct", "Lung_Michigan_collapsed_symbols.gct" and "Lung_Boston_collapsed_symbols.gct" beforehand downloaded from the GSEA site. Then call demo(diabetes) and demo(boston) or open the files and execute the code manually. The processing takes hours. 
}

\keyword{datasets, analysis}
