\name{cnSetSeed}
\alias{cnSetSeed}

\title{Random Generator Seed}

\description{Sets a seed for the random number generator.}

\usage{
cnSetSeed(seed)
}

\arguments{
\item{seed}{an \code{integer}}
}

\details{Setting a fixed seed before any stochastic function guaratees repeated results.
}

\value{NA}

\author{N. Balov}

\keyword{methods}
\keyword{graphs}
