\name{CPDAG-class}
\docType{class}
\alias{CPDAG}
\alias{CPDAG-class}

\title{Class CPDAG}
\description{Base class implementing Complete Partially Directed Acyclic Graphs (CPDAGs)}

\section{Slots}{

\describe{
	\item{\code{numnodes}:}{an integer, the number of nodes}
	\item{\code{nodes}:}{a \code{vector} of node names}
	\item{\code{edges}:}{a \code{list} of graph edges}
  }
}

\author{N. Balov}

\seealso{\code{\link{dag2cpdag}}
}

\keyword{classes}
\keyword{graphs}

