\name{dag2cpdag-method}
\docType{methods}
\alias{dag2cpdag}
\alias{dag2cpdag,catNetwork-method}

\title{Complete Network Representation}

\description{Generate the complete graphical structure for a \code{catNetwork} object.}

\usage{
dag2cpdag(object)
}

\arguments{
\item{object}{a \code{catNetwork} object}
}

\value{A non-DAG \code{catNetwork} object.}

\author{N. Balov}

\keyword{methods}
\keyword{graphs}
