\name{novartis}
\docType{data}
\alias{novartis}

\title{Novartis multi-tissue data}
\description{
  Consensus Clustering: A re-sampling-based method for class discovery and visualization of gene expression microarray data
}

\usage{data(novartis)}
\format{A matrix containing 105 observations on 500 genes.}
\source{"http://www.broadinstitute.org/cgi-bin/cancer/datasets.cgi"}
\keyword{datasets}
